requirements.txt
asyncio==3.4.3
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.3
numpy==1.16.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1

psycopg2-binary==2.7.7
django==2.1.7
channels==2.1.7
channels_redis==2.3.3
selenium==3.141.0
geopy==1.18.1
gunicorn==19.9.0
whitenoise==4.1.2